+++ cmd/scheduler/app/options/options.go
@@ -14,7 +14,7 @@ import (
 
 const (
 	defaultSchedulerName               = "kai-scheduler"
-	defaultResourceReservationAppLabel = "runai-reservation"
+	defaultResourceReservationAppLabel = "kai-resource-reservation"
 	defaultMetricsNamespace            = "kai"
 	defaultSchedulerPeriod             = time.Second
 	defaultStalenessGracePeriod        = 60 * time.Second