+++ test/e2e/modules/resources/rd/pytorch/pytorch.go
@@ -73,7 +73,7 @@ func GetPodTemplate(queueName, matchLabelValue string) corev1.PodTemplateSpec {
 		ObjectMeta: metav1.ObjectMeta{
 			Labels: map[string]string{
 				constants.AppLabelName: matchLabelValue,
-				"runai/queue":          queueName,
+				"kai.scheduler/queue":  queueName,
 			},
 		},
 		Spec: corev1.PodSpec{