+++ test/e2e/modules/resources/rd/replicaset.go
@@ -42,7 +42,7 @@ func CreateReplicasetObject(namespace, queueName string) *v1.ReplicaSet {
 					Labels: map[string]string{
 						constants.AppLabelName: "engine-e2e",
 						ReplicaSetAppLabel:     matchLabelValue,
-						"runai/queue":          queueName,
+						"kai.scheduler/queue":  queueName,
 					},
 				},
 				Spec: v12.PodSpec{